+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Dropped support for using title attribute to add link modifiers and icons.
 - Added support for link modifiers and icons through WP Menu Builder 'classes' input.
 	- Icon support is for Font Awesome 4/5 and Glyphicons icons.
+- Added unit tests for the `fallback` method.
 
 ## [3.0.0]